feat: add implementations of main backend handlers: creating, getting services&streams

+++ backend/internal/domain/service.go
@@ -2,5 +2,5 @@ package domain
 
 type Service struct {
 	Name string `json:"name"`
-	Port uint16 `json:"port"`
+	Port int32  `json:"port"`
 }
